Rename measures to restrictions
