+++ packages/cms/schemas/restrictions/restriction-category.js
@@ -1,20 +1,20 @@
 export default {
-  title: "Categorie",
-  name: "measureCategory",
+  title: "Restrictie Categorie",
+  name: "restrictionCategory",
   type: "object",
   fields: [
     {
       title: "Titel van groep",
-      description: "Hoe noem je deze groep?",
+      description: "Hoe noem je deze groep maatregelen?",
       name: "title",
       type: "localeString",
     },
     {
       title: "Maatregelen",
       description: "Per groep bestaat er een lijst maatregelen",
-      name: "measures",
+      name: "restrictions",
       type: "array",
-      of: [{ type: "measureGroup" }],
+      of: [{ type: "restrictionGroup" }],
     },
   ],
   preview: {
